% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/function.R
\name{Function}
\alias{Function}
\title{The Function Monad}
\description{
The \code{function} built-in type is a functor with function composition as \code{\link[=fmap]{fmap()}}.
}
\details{
It's also possible to define \code{\link[=bind]{bind()}} for Function in a way that satisfies
the \link[=monad-laws]{monad laws}: given the output of the first function with
some inputs, the second function returns the next function to apply to the
same inputs. And indeed that definition is provided, however its practical
uses may be limited.
}
\examples{
# The fmap operator corresponds to function composition.
p1log <- identity \%>>\% `+`(1) \%>>\% log()
p1log(0.5) == log1p(0.5)
}
\seealso{
Other monads: 
\code{\link{List}},
\code{\link{Maybe}}
}
\concept{monads}
