% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/functor-laws.R
\name{functor-laws}
\alias{functor-laws}
\alias{expect_preserve_identity}
\alias{expect_preserve_composition}
\title{Functor Laws}
\usage{
expect_preserve_identity(m)

expect_preserve_composition(f, g, m)
}
\arguments{
\item{m}{A functor object.}

\item{f, g}{Functions.}
}
\description{
Classes implementing \code{\link[=fmap]{fmap()}} are expected to satisfy two functor laws.
Functions to test if the laws hold are provided for use with the \code{testthat}
package.
}
\details{
The Haskell functor laws have been translated to R in the expectations as
follows:

\describe{
\item{Preservation of identity:}{\code{m \%>>\% identity} is equal to \code{m |> identity()}.}
\item{Preservation of composition:}{\code{m \%>>\% (f \%.\% g)} is equal to \code{m \%>>\% g \%>>\% f}.}
}

Where above \verb{\%.\%} denotes function composition.
}
\references{
\url{https://wiki.haskell.org/Functor#Functor_Laws}
}
\seealso{
Other implementation laws: 
\code{\link{monad-laws}}
}
\concept{implementation laws}
