% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/maybe.R, R/result.R
\name{maybe}
\alias{maybe}
\alias{some}
\alias{none}
\alias{ok}
\alias{err}
\title{The Maybe Monad}
\usage{
some(value)

none()

ok(value)

err(value)
}
\arguments{
\item{value}{An object.}
}
\description{
The Maybe monad represents the presence or absence of a value.
}
\examples{
# fmap() transforms the wrapped value, if there is one.
some(1) \%>>\% `+`(1)
none() \%>>\% `+`(1)

# bind() should be used with functions that return a maybe.
some(1) \%>-\% function(x) none()
some(1) \%>>\% function(x) none() # Probably not what you wanted.
}
