% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/maybe.R
\name{maybe}
\alias{maybe}
\title{The Maybe Monad}
\description{
The package \href{https://armcn.github.io/maybe/}{maybe} implements the Maybe
monad. Here method implementations for \code{\link[=fmap]{fmap()}}, \code{\link[=bind]{bind()}} and \code{\link[=join]{join()}}
are provided for the \code{maybe} S3 class from the package. The methods are
simple wrappers for the corresponding functions in maybe.
}
\examples{
\dontshow{if (rlang::is_installed("maybe")) (if (getRversion() >= "3.4") withAutoprint else force)(\{ # examplesIf}
# The fmap operator corresponds to maybe::maybe_map().
maybe::just(1) \%>>\% `+`(1)
maybe::nothing() \%>>\% `+`(1)

# The bind operator corresponds to maybe::and_then().
maybe::just(1) \%>-\% \(x) maybe::just(x + 1)
maybe::just(1) \%>-\% \(x) maybe::nothing()
maybe::nothing() \%>-\% \(x) maybe::just(1)
\dontshow{\}) # examplesIf}
}
\seealso{
\code{\link[maybe:maybe_map]{maybe::maybe_map()}} which implements \code{\link[=fmap]{fmap()}} for \code{maybe} values.

\code{\link[maybe:and_then]{maybe::and_then()}} which implements \code{\link[=bind]{bind()}} for \code{maybe} values.
}
