% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/monad-package.R
\docType{package}
\name{monad-package}
\alias{monad-package}
\title{monad: Monad Operators and Generics}
\description{
Operators `%>>%` and `%>-%` let you compose generic monadic function pipelines based on implementing the S7 generics `fmap()` and `bind()`. Methods are provided for the built-in `list` type and the `maybe` class from the 'maybe' package. The concepts are modelled directly after the Monad typeclass in Haskell, but adapted for idiomatic use in R.
}
\seealso{
Useful links:
\itemize{
  \item \url{https://github.com/mikmart/monad}
  \item \url{https://mikmart.github.io/monad/}
  \item Report bugs at \url{https://github.com/mikmart/monad/issues}
}

}
\author{
\strong{Maintainer}: Mikko Marttila \email{mikkmart@protonmail.com}

}
\keyword{internal}
