% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/monad-package.R
\docType{package}
\name{monad-package}
\alias{monad}
\alias{monad-package}
\title{monad: Monad Operators and Generics}
\description{
What the package does (one paragraph).
}
\author{
\strong{Maintainer}: Mikko Marttila \email{mikkmart@protonmail.com}

}
\keyword{internal}
